zmongo_toolbag/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zmongo_toolbag"
version = "0.1.5"
description = "Seamless MongoDB retrieval operations using OpenAI GPT and local LLaMA models."
readme = "README.md"
license = { file = "LICENSE.md" }
authors = [
    { name = "CentralFloridaAttorney" }
]
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "pymongo",
    "motor",
    "pandas",
    "openai",
    "beautifulsoup4",
    "nest_asyncio",
    "nbformat",
    "ace-tools",
    "matplotlib",
    "opencv-python",
    "mtcnn",
    "langchain",
    "pytest",
    "pytest-asyncio",
    "astor",
    "numpy",
    "tiktoken",
    "google-generativeai"
]

[project.optional-dependencies]
dev = [
    "black>=23.0",
    "isort>=5.10.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0",
]
test = [
    "pytest",
    "pytest-asyncio",
    "coverage[toml]",
    "pytest-cov"
]
docs = [
    "mkdocs",
    "mkdocstrings[python]",
    "markdown-include"
]
notebooks = [
    "jupyterlab",
    "ipykernel",
    "ipywidgets"
]
lint = [
    "flake8",
    "pylint",
    "bandit"
]

[project.urls]
Homepage = "https://github.com/CentralFloridaAttorney/zmongo_retriever"
Repository = "https://github.com/CentralFloridaAttorney/zmongo_retriever"
Issues = "https://github.com/CentralFloridaAttorney/zmongo_retriever/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["zmongo_toolbag*"]
